Expose many more configuration variables as facts
These variables will be needed if we want to manage the
puppet.conf file on all operating systems but keep local
overrides for the most important parameters.  If we didn't
preserve those variables they would fall back to their
built-in default value.  The negative side is that all of
these values will become sticky and cannot be changed for
just a single Puppet run.