Simple-minded management of puppet.conf
Include puppet::config and individual nested classes to manage the
respective sections of the puppet.conf file.  You can also include
the puppet::agent and/or puppet::master classes to achieve the same
effect plus more management of Puppet agent/master aspects in the
future.

Caveat: What's managed in the configuration and what's not is
currently very much hard-coded and the [master] section is actually
static and only works for me.